Abstract duplicate Edit methods to separate class and add defensive programming to Schedule
